[Wrench Est.]: create a base class for estimators that requiring actuator measurements
